+++ .github/workflows/ci.yml
@@ -223,6 +223,9 @@ jobs:
           cd ../fetch
           cargo publish
           sleep 30
+          cd ../crypto
+          cargo publish
+          sleep 30
           cd ../../cli
           sleep 30
           cargo publish
