+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
   "test_util",
   "op_crates/fetch",
   "op_crates/web",
+  "op_crates/crypto"
 ]
 exclude = [
   "std/hash/_wasm"
