+++ cli/Cargo.toml
@@ -20,6 +20,7 @@ harness = false
 path = "./bench/main.rs"
 
 [build-dependencies]
+deno_crypto = { path = "../op_crates/crypto", version = "0.1.1" }
 deno_core = { path = "../core", version = "0.67.0" }
 deno_web = { path = "../op_crates/web", version = "0.18.0" }
 deno_fetch = { path = "../op_crates/fetch", version = "0.10.0" }
@@ -31,6 +32,7 @@ winres = "0.1.11"
 winapi = "0.3.9"
 
 [dependencies]
+deno_crypto = { path = "../op_crates/crypto", version = "0.1.1" }
 deno_core = { path = "../core", version = "0.67.0" }
 deno_doc = "0.1.15"
 deno_lint = "0.2.9"
@@ -55,7 +57,6 @@ libc = "0.2.77"
 log = "0.4.11"
 env_logger = "0.7.1"
 notify = "5.0.0-pre.3"
-rand = "0.7.3"
 regex = "1.3.9"
 ring = "0.16.15"
 rustyline = { version = "6.3.0", default-features = false }
