+++ cli/ops/crypto.rs
@@ -1,0 +1,14 @@
+// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+use deno_crypto::op_get_random_values;
+use deno_crypto::rand::rngs::StdRng;
+use deno_crypto::rand::SeedableRng;
+
+pub fn init(rt: &mut deno_core::JsRuntime, maybe_seed: Option<u64>) {
+  if let Some(seed) = maybe_seed {
+    let rng = StdRng::seed_from_u64(seed);
+    let op_state = rt.op_state();
+    let mut state = op_state.borrow_mut();
+    state.put::<StdRng>(rng);
+  }
+  super::reg_json_sync(rt, "op_get_random_values", op_get_random_values);
+}
