+++ cli/ops/fs.rs
@@ -13,8 +13,8 @@ use deno_core::serde_json::Value;
 use deno_core::BufVec;
 use deno_core::OpState;
 use deno_core::ZeroCopyBuf;
-use rand::thread_rng;
-use rand::Rng;
+use deno_crypto::rand::thread_rng;
+use deno_crypto::rand::Rng;
 use serde::Deserialize;
 use std::cell::RefCell;
 use std::convert::From;
