+++ op_crates/crypto/01_crypto.js
@@ -2,11 +2,11 @@
 
 ((window) => {
   const core = window.Deno.core;
-  const { assert } = window.__bootstrap.util;
-
   function getRandomValues(typedArray) {
-    assert(typedArray !== null, "Input must not be null");
-    assert(typedArray.length <= 65536, "Input must not be longer than 65536");
+    if (typedArray == null) throw new Error("Input must not be null");
+    if (typedArray.length > 65536) {
+      throw new Error("Input must not be longer than 65536");
+    }
     const ui8 = new Uint8Array(
       typedArray.buffer,
       typedArray.byteOffset,
@@ -15,7 +15,10 @@
     core.jsonOpSync("op_get_random_values", {}, ui8);
     return typedArray;
   }
-
+  window.crypto = {
+    getRandomValues,
+  };
+  window.__bootstrap = window.__bootstrap || {};
   window.__bootstrap.crypto = {
     getRandomValues,
   };
