+++ op_crates/crypto/Cargo.toml
@@ -1,0 +1,19 @@
+# Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+
+[package]
+name = "deno_crypto"
+version = "0.1.1"
+edition = "2018"
+description = "Collection of WebCrypto APIs"
+authors = ["the Deno authors"]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/denoland/deno"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+deno_core = { version = "0.67.0", path = "../../core" }
+rand = "0.7.3"
+
