+++ op_crates/crypto/README.md
@@ -1,0 +1,3 @@
+# deno crypto
+
+Op crate that implements crypto functions.
