+++ op_crates/crypto/lib.rs
@@ -1,26 +1,31 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 
+#![deny(warnings)]
+
 use deno_core::error::AnyError;
 use deno_core::serde_json::json;
 use deno_core::serde_json::Value;
+use deno_core::JsRuntime;
 use deno_core::OpState;
 use deno_core::ZeroCopyBuf;
 use rand::rngs::StdRng;
 use rand::thread_rng;
 use rand::Rng;
-use rand::SeedableRng;
 
-pub fn init(rt: &mut deno_core::JsRuntime, maybe_seed: Option<u64>) {
-  if let Some(seed) = maybe_seed {
-    let rng = StdRng::seed_from_u64(seed);
-    let op_state = rt.op_state();
-    let mut state = op_state.borrow_mut();
-    state.put::<StdRng>(rng);
+pub use rand; // Re-export rand
+
+/// Execute this crates' JS source files.
+pub fn init(isolate: &mut JsRuntime) {
+  let files = vec![(
+    "deno:op_crates/crypto/01_crypto.js",
+    include_str!("01_crypto.js"),
+  )];
+  for (url, source_code) in files {
+    isolate.execute(url, source_code).unwrap();
   }
-  super::reg_json_sync(rt, "op_get_random_values", op_get_random_values);
 }
 
-fn op_get_random_values(
+pub fn op_get_random_values(
   state: &mut OpState,
   _args: Value,
   zero_copy: &mut [ZeroCopyBuf],
